+++ src/shared/model/svg.ts
